+++ backend/utils/types.ts
@@ -1,0 +1,12 @@
+export interface UserRow {
+	id: number;
+	email: string;
+	password: string;
+	username: string;
+}
+
+export interface UserResult {
+	idUser: number;
+	username: string;
+	email: string;
+}
