+++ frontend/src/utils/types.ts
@@ -10,3 +10,15 @@ export interface Ticket {
 	idUser: number;
 	username: string;
 }
+
+export interface RegisterData {
+	username: string;
+	email: string;
+	password: string;
+	role: string;
+}
+
+export interface LoginData {
+	email: string;
+	password: string;
+}
