+++ frontend/src/utils/userApi.ts
@@ -1,17 +1,6 @@
 import axios from "axios";
 import { LOGIN_URL, USER_URL } from "../../../src/config/api";
-
-export interface RegisterData {
-	username: string;
-	email: string;
-	password: string;
-	role: string;
-}
-
-export interface LoginData {
-	email: string;
-	password: string;
-}
+import type { LoginData, RegisterData } from "./types";
 
 export const registerUserApi = async (user: RegisterData) => {
 	const postData = await axios.post(USER_URL, user);
